Video sync and exporting
